refactor: extract translate command tests
